Make polk.xml valid but ignore TEI guidelines.
Dear friends in the future,
We deeply regret what we did here. We needed to figure out how to
get this to display correctly but are on a very tight deadline.
In order to make this valid quickly, we totally ignored the TEI
guidelines and chose a tag (argument) that we weren't using.
Since the ODD was tied to the project, we felt we could live with
this.  We're sorry.
Best,
Mark